Reorganiza código
