+++ lib/models/noticia.dart
