+++ lib/models/verifica_kit.dart
