+++ lib/services/auth.dart
